Major overhaul of the current cli tool. It now accepts 3 different inputs that work regardless of ordering.
We now have versioning for the cli tool that looks at the outermost version.  This is used along with the option '-v' or '--version' to get the current version of the program. Another option is '-h' or '--help' to get info about how to run the program and what options are available. The final working option is '-p' or '--port' where you can specify port to be used.
We probably should alter this from beeing hardcoded as it is now to instead have somme way (maybe a struct) that has the information so that we always show all available options and never more or less.